requirements.txt
pytest-playwright==0.7.0
# playwright install
fastapi[standard]==0.115.11
pyhumps==3.8.0
PyJWT==2.10.1
pytz==2024.2
sqlmodel==0.0.22